+++ tests/Features/Fixtures/ThreadSearchableTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Features\Fixtures;
+
+trait ThreadSearchableTrait
+{
+    public function toSearchableArray()
+    {
+        return [
+            'something' => 99,
+        ];
+    }
+}